requirements.txt
django
djangorestframework
psycopg2
Django-cors-headers
gunicorn
django-heroku